+++ runner/android_junit_runner/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * Exceptions during `@AfterClass` were not being reported via `InstrumentationResultPrinter`.
 * Exceptions arising in AndroidJUnitRunner.buildRequest are now handled.
+* Assumption failures during a ClassRule or BeforeClass are now reported more consistently via `InstrumentationResultPrinter`
 
 **New Features**